@@ -8,46 +8,45 @@
 
 namespace simialbi\yii2\elfinder;
 
-
 class ElFinderConfigurationMySQL extends ElFinderConfiguration
 {
     /**
      * @var string Volume driver. Set storage engine for current root, can be one of LocalFileSystem, MySQL, FTP
      */
-    public $driver = ElFinder::DRIVER_MYSQL;
+    public string $driver = ElFinder::DRIVER_MYSQL;
 
     /**
      * @var string MySQL host name
      */
-    public $host = 'localhost';
+    public string $host = 'localhost';
 
     /**
      * @var string|null Path to MySQL sock file
      */
-    public $socket = null;
+    public ?string $socket = null;
 
     /**
      * @var integer|null MySQL connection port (TCP)
      */
-    public $port = null;
+    public ?int $port = null;
 
     /**
      * @var string MySQL user name
      */
-    public $user = '';
+    public string $user = '';
 
     /**
      * @var string MySQL password
      */
-    public $pass = '';
+    public string $pass = '';
 
     /**
      * @var string MySQL database name
      */
-    public $db = '';
+    public string $db = '';
 
     /**
      * @var string MySQL files table name
      */
-    public $files_table = 'elfinder_file';
+    public string $files_table = 'elfinder_file';
 }
